# Conflicts:
#	src/Js/src/globals.ts
#	src/Js/src/inscore.ts
#	src/Js/src/model/IRect.ts
#	src/javascript/src/controller/IGlue.ts
#	src/javascript/src/events/documentEvents.ts
#	src/javascript/src/model/IObject.ts
#	src/javascript/src/model/TILoader.ts
#	src/javascript/src/vhtmlview/VHtmlSvg.ts
#	src/javascript/testhtml/tests1.inscore


Former-commit-id: 7e5ad9fd85058044bd5ac1026fe47704cd62b056

+++ tools/IBundle/src/Parsing/ParsedData.cpp
@@ -198,7 +198,7 @@ std::string ParsedData::generateScript(std::string name) const
 	return script.str();
 }
 
-void ParsedData::simplifyPaths(int charToDelete)
+void ParsedData::simplifyPaths(size_t charToDelete)
 {
 	if(mainScript.empty() && fScripts.empty() && fRessources.empty())
 		return;
@@ -248,13 +248,13 @@ void ParsedData::simplifyPaths()
 
 	//  -- search for common path trunk --
 	std::vector<std::string> trunk = splitPath(fRessources.size()? fRessources.begin()->first: fScripts.begin()->first);
-	int commonPath = trunk.size()-1;
+	size_t commonPath = trunk.size()-1;
 
 	for(auto it = fRessources.cbegin(); it != fRessources.cend(); it++){
 		std::vector<std::string> path = splitPath(it->first);
 
-		int i=0;
-		while ( i < commonPath && i < (int)path.size()-1 ) {
+		size_t i=0;
+		while ( i < commonPath && i < path.size()-1 ) {
 			if(path.at(i)!=trunk.at(i)){
 				break;
 			}
@@ -266,8 +266,8 @@ void ParsedData::simplifyPaths()
 	for(auto it = fScripts.cbegin(); it != fScripts.cend(); it++){
 		std::vector<std::string> path = splitPath(it->first);
 
-		int i=0;
-		while ( i < commonPath && i < (int)path.size()-1 ) {
+		size_t i=0;
+		while ( i < commonPath && i < path.size()-1 ) {
 			if(path.at(i)!=trunk.at(i))
 				break;
 			i++;
@@ -280,7 +280,7 @@ void ParsedData::simplifyPaths()
 
 	std::string trunkPath = "/";
 
-	for (int i = 0; i < commonPath; ++i)
+	for (size_t i = 0; i < commonPath; ++i)
 		trunkPath += trunk.at(i)+"/";
 
 	size_t removable = trunkPath.size();
